small change
so, all we need right now for the presentations is the top restaurants page finished (this means that we change add links to each of the titles that lead to the restaurants webpage).
We also need a show and tell day java feature and maybe a few more things. We also need to go over how we're going to do the presentation